\name{IPO-package}
\alias{IPO-package}
\alias{IPO}
\docType{package}
\title{Automated Optimization of Untargeted Metabolomics LC-MS Data Processing}
\description{
IPO provides a framework for parameter optimization for the software package XCMS.
It provides optimisation of peak picking parameters by using natural, stable 13C isotopes. Retention time correction is optimized by minimizing the relative retention time differences 
within features and grouping parameters are optimized by maximizing the number of features 
showing exactly one peak from each injection of a pooled sample.
}
\details{
An overview of how to use the package, including the most important functions
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>

}
\references{
Literature or other references for background information
}

# Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
# directory ~~
\keyword{ package }
\seealso{
# ~~ Optional links to other man pages, e.g. ~~
#\code{\link[<pkg>:<pkg>-package]{<pkg>}}
\code{\link[xcms:xcmsSet]{xcms}}
}
\examples{
# simple examples of the most important functions 
}
